Oops.  Use "unsigned long" instead of "int" for the intermediate variables in wide-character conversions, since it's guaranteed to be large enough. Thanks to: Andrey Chernov
git-svn-id: svn+ssh://svn.freebsd.org/base/head@133304 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ lib/libarchive/archive_read_support_format_tar.c
@@ -1519,7 +1519,7 @@ static size_t
 UTF8_mbrtowc(wchar_t *pwc, const char *s, size_t n)
 {
         int ch, i, len, mask;
-        int lbound, wch;
+        unsigned long lbound, wch;
 
         if (s == NULL)
                 /* Reset to initial shift state (no-op) */
